fix: refactored routing
